+++ components/DocMenu/react/_DocLink.tsx
@@ -23,6 +23,8 @@ export const DocLink: React.FC<DocLinkProps> = ({
     if (item.active) {
       onActive?.(activeLIRef?.current?.offsetTop || 0)
     }
+    // supposed to happen only on mount
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
   return (
@@ -46,7 +48,7 @@ export const DocLink: React.FC<DocLinkProps> = ({
               'absolute w-[4px] z-10 top-[10%] h-[80%] rounded-full hidden',
               {
                 'group-hover:block bg-gray-300': !item.active && collapsible,
-              }
+              },
             )}
             style={{
               left: `-${18.5 + depth * 8}px`,