+++ components/StatusIcon/constants/src/outline-imports.ts
@@ -41,6 +41,8 @@ import {
   cyStatusRunningOutlineX24,
   cyStatusRunningSimpleX4,
   cyStatusRunningSimpleX8,
+  cyStatusFailingOutlineX16,
+  cyStatusFailingOutlineX24,
   cyStatusPlaceholderSolidX12,
   cyStatusPlaceholderSolidX16,
   cyStatusPlaceholderSolidX24,
@@ -55,6 +57,13 @@ export const statuses: Record<string, IconSet> = {
     size12Icon: cyStatusRunningOutlineX12,
     size16Icon: cyStatusRunningOutlineX16,
     size24Icon: cyStatusRunningOutlineX24,
+  },
+  failing: {
+    size4Icon: cyStatusRunningSimpleX4,
+    size8Icon: cyStatusRunningSimpleX8,
+    size12Icon: cyStatusRunningOutlineX12,
+    size16Icon: cyStatusFailingOutlineX16,
+    size24Icon: cyStatusFailingOutlineX24,
   },
   passed: {
     size4Icon: cyStatusPassedSimpleX4,